+++ kinto/plugins/history/listener.py
@@ -1,5 +1,4 @@
 from kinto.core.utils import instance_uri
-import copy
 from datetime import datetime
 
 
@@ -9,20 +8,13 @@ def on_resource_changed(event):
     ``history`` resource. The entries are served as read-only in the
     :mod:`kinto.plugins.history.views` module.
     """
-    payload = copy.deepcopy(event.payload)
+    payload = event.payload
     resource_name = payload['resource_name']
     event_uri = payload['uri']
 
-    bucket_id = payload.pop('bucket_id')
-    bucket_uri = instance_uri(event.request, 'bucket', id=bucket_id)
-    collection_id = None
+    bucket_id = None
+    bucket_uri = None
     collection_uri = None
-    if 'collection_id' in payload:
-        collection_id = payload['collection_id']
-        collection_uri = instance_uri(event.request,
-                                      'collection',
-                                      bucket_id=bucket_id,
-                                      id=collection_id)
 
     storage = event.request.registry.storage
     permission = event.request.registry.permission
@@ -31,12 +23,30 @@ def on_resource_changed(event):
     for impacted in event.impacted_records:
         target = impacted['new']
         obj_id = target['id']
+
+        try:
+            bucket_id = payload['bucket_id']
+        except KeyError:
+            # e.g. DELETE /buckets
+            bucket_id = obj_id
+        bucket_uri = instance_uri(event.request, 'bucket', id=bucket_id)
+
+        if 'collection_id' in payload:
+            collection_id = payload['collection_id']
+            collection_uri = instance_uri(event.request,
+                                          'collection',
+                                          bucket_id=bucket_id,
+                                          id=collection_id)
+
         # On POST .../records, the URI does not contain the newly created
-        # record id. Make sure it does:
-        if event_uri.endswith(obj_id):
-            uri = event_uri
+        # record id.
+        parts = event_uri.split('/')
+        if resource_name in parts[-1]:
+            parts.append(obj_id)
         else:
-            uri = event_uri + '/' + obj_id
+            # Make sure the id is correct on grouped events.
+            parts[-1] = obj_id
+        uri = '/'.join(parts)
         targets.append((uri, target))
 
     # Prepare a list of object ids to be fetched from permission backend,
@@ -67,6 +77,7 @@ def on_resource_changed(event):
         # Prepare the history entry attributes.
         perms = {k: list(v) for k, v in perms_by_object_id[uri].items()}
         eventattrs = dict(**payload)
+        eventattrs.pop('bucket_id', None)
         eventattrs.setdefault('%s_id' % resource_name, obj_id)
         eventattrs['uri'] = uri
         attrs = dict(date=datetime.now().isoformat(),